requirements.txt
fastapi==0.68.0
python-multipart==0.0.5
google-cloud-vision==3.4.4
pillow==10.0.1
mangum==0.17.0 